requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
telethon==1.34.0
